dev-requirements.txt
asgiref==3.7.2
build==1.0.0
click==8.1.7
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
importlib-metadata==6.8.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
pip-tools==7.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
